improved the swagger documentation & other changes

+++ customers/views.py
@@ -13,13 +13,14 @@ from .serializers import CustomerSerializer, CustomerUpdateSerializer
 # Create your views here.
 class CustomerCreate(APIView):
 
+    permission_classes = [IsAdminUser]
+
     def get_permissions(self):
         if self.request.method == "POST":
             permission_classes = [AllowAny]
-        else:
-            permission_classes = [IsAdminUser]
-        return [permission() for permission in permission_classes]
-
+            return [permission() for permission in permission_classes]
+        return super().get_permissions()
+        
     def get(self, request, *args, **kwargs):
         customers = Customer.objects.all()
         serializer = CustomerSerializer(customers, many=True)
@@ -38,13 +39,6 @@ class CustomerInstance(APIView):
 
     permission_classes = [CustomerOnly]
 
-    def get_permissions(self):
-        if self.request.method in ["PUT", "PATCH"]:
-            permission_classes = [CustomerOnly]
-        else:
-            permission_classes = [IsAdminUser | CustomerOnly]
-        return [permission() for permission in permission_classes]
-
     def get_object(self, pk):
         try:
             return Customer.objects.get(pk=pk)
@@ -52,13 +46,16 @@ class CustomerInstance(APIView):
             raise exceptions.NotFound({"error": "Customer with supplied ID doesn't exist."})
 
     def get(self, request, pk, *args, **kwargs):
+        self.check_permissions(request)
         customer = self.get_object(pk=pk)
+        self.check_object_permissions(request, obj=customer)
         serializer = CustomerSerializer(customer)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     @swagger_auto_schema(request_body=CustomerUpdateSerializer)
     def put(self, request, pk, *args, **kwargs):
         customer = self.get_object(pk=pk)
+        self.check_object_permissions(request, obj=customer)
         serializer = CustomerUpdateSerializer(customer, data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -67,6 +64,7 @@ class CustomerInstance(APIView):
 
     def delete(self, request, pk, *args, **kwargs):
         customer = self.get_object(pk=pk)
+        self.check_object_permissions(request, obj=customer)
         customer.delete()
         return Response(
             {"message": "Customer has been deleted"}, status=status.HTTP_204_NO_CONTENT
